Add show profile info
some auth fixes
some save challenge in DB fixes

+++ app/src/main/java/com/myrungo/rungo/run/RunView.kt
@@ -11,7 +11,7 @@ interface RunView : MvpView {
     fun showSkin(skin: CatView.Skins)
     fun showMap(show: Boolean)
     fun showTime(curTime: String, challengeTime: String)
-    fun showSpeed(curSpeed: Float, avgSpeed: Float)
+    fun showSpeed(curSpeed: Double, avgSpeed: Double)
     fun showDistance(curDistance: String, challengeDistance: String)
     fun run(isRun: Boolean)
